+++ front/src/typings/index.ts
@@ -4,3 +4,4 @@ export * from './account';
 export * from './resource';
 export * from './service';
 export * from './business';
+export * from './loadbalancer';
