+++ front/src/typings/loadbalancer.ts
@@ -1,0 +1,1 @@
+export type IOriginPage = 'lb' | 'listener' | 'domain';
