+++ front/src/views/business/load-balancer/clb-view/components/AddOrUpdateDomainSideslider/index.scss
@@ -1,0 +1,22 @@
+.domain-sideslider {
+  .readonly-info {
+    display: flex;
+    align-items: center;
+    height: 40px;
+
+    .label {
+      margin-right: 5px;
+      width: 90px;
+      text-align: right;
+    }
+
+    .value {
+      margin-left: 10px;
+      color: #313238;
+    }
+
+    &:last-of-type {
+      margin-bottom: 12px;
+    }
+  }
+}
