+++ front/src/views/business/load-balancer/clb-view/components/AddOrUpdateListenerSideslider/index.scss
@@ -1,0 +1,11 @@
+.listener-sideslider-container {
+  .form-item-tips {
+    height: 20px;
+    line-height: 20px;
+    font-size: 12px;
+    color: #979ba5;
+  }
+  .recommend-tag {
+    margin-left: 4px !important;
+  }
+}
