+++ front/src/views/business/load-balancer/clb-view/specific-clb-manager/listener-list/index.scss
@@ -1,15 +1,3 @@
-.listener-sideslider-container {
-  .form-item-tips {
-    height: 20px;
-    line-height: 20px;
-    font-size: 12px;
-    color: #979ba5;
-  }
-  .recommend-tag {
-    margin-left: 4px !important;
-  }
-}
-
 .batch-delete-listener-dialog {
   .bkhcm-icon-minus-circle-shape {
     font-size: 14px;
