+++ front/src/views/business/load-balancer/clb-view/specific-listener-manager/domain-list/index.scss
@@ -25,26 +25,3 @@
   color: #c4c6cc;
   cursor: pointer;
 }
-
-.domain-sideslider {
-  .readonly-info {
-    display: flex;
-    align-items: center;
-    height: 40px;
-
-    .label {
-      margin-right: 5px;
-      width: 80px;
-      text-align: right;
-    }
-
-    .value {
-      margin-left: 10px;
-      color: #313238;
-    }
-
-    &:last-of-type {
-      margin-bottom: 12px;
-    }
-  }
-}
